requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
et-xmlfile==1.1.0
fake-headers==1.0.2
frozenlist==1.4.1
html5lib==1.1
idna==3.6
iniconfig==2.0.0
loguru==0.7.2
multidict==6.0.5
mypy==1.9.0
mypy-extensions==1.0.0
openpyxl==3.1.2
packaging==24.0
pluggy==1.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyPDF2==3.0.1
pytest==8.1.1
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
types-requests==2.31.0.20240311
typing_extensions==4.10.0
urllib3==2.2.1
webencodings==0.5.1
win32-setctime==1.1.0
yarl==1.9.4
